Implement agent.task_queue
The purpose of this task_queue is to allow multiple agents to simultaneously set the same print task A as their next possible task, provided that these agents must not be printing any other tasks at the moment and the target task A has not been printed. During agent movement, once task A has been reached by an agent at its endpoint and started to print, the other agents will dynamically and automatically exclude task A from their task_queue list and select another task in agent.task_queue to print. Whenever an agent finishes printing a target task, it empties the task_queue and recalculates the optimal two potential tasks to be added to agent.task_queue according to the auction algorithm.